pyproject.toml
[tool.poetry]
name = "pengle"
version = "0.1.0"
description = ""
authors = ["pesuchin <gp11a110@oecu.jp>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
mlflow = "^1.8.0"
feather-format = "^0.4.1"
sklearn = "^0.0"
category_encoders = "^2.2.2"
lightgbm = "^2.3.1"
optuna = "^1.4.0"

[tool.poetry.dev-dependencies]
yapf = "^0.30.0"
pylint = "^2.5.0"
isort = "^4.3.21"
autopep8 = "^1.5.2"
pytest = "^5.4.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
